pyproject.toml
[project]
name = "comfyui-tosvg"
description = "This project converts raster images into SVG format using the [a/VTracer](https://github.com/visioncortex/vtracer) library. It's a handy tool for designers and developers who need to work with vector graphics programmatically."
version = "1.0.1"
license = { file = "LICENSE" }
dependencies = ["numpy", "Pillow", "torch", "vtracer"]

[project.urls]
Repository = "https://github.com/Yanick112/ComfyUI-ToSVG"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "Yanick112"
DisplayName = "ComfyUI-ToSVG"
Icon = ""
